+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Update to pinky-swear 3.0.0 (properly handle chaining result promises)
 * `Confirmation` can no longer hold an error
 * `PublisherConfirm` now returns a proper `Result<Confirmation>`
+* Connecting now returns a `CloseOnDrop<Connection>`
+* Creating a channel now returns a `CloseOnDrop<Channel>`
 
 #### Features